feat : 2Stage 좀비 및 아이템 스폰 포인트 지정 완료

+++ Assets/02.Scripts/01.Network/NetworkManager.cs
@@ -152,7 +152,7 @@ public class NetworkManager : MonoBehaviourPunCallbacks {
     public void CreatePlayer() {
         GameObject player = PhotonNetwork.Instantiate("PlayerPrefab", transform.position, Quaternion.identity);
         GameObject[] playerSpawnPoint = GameObject.FindGameObjectsWithTag("PlayerSpawnPoint");
-        int index = (player.GetComponent<PhotonView>().ViewID / 1000) % 4;
+        int index = ((int)player.GetComponent<PhotonView>().ViewID / 1000) % 4;
         switch (index) {
             case 0:
                 player.transform.position = playerSpawnPoint[0].transform.position;
